03_lens_tester_gui/requirements.txt
pyyaml
pyqt5
numpy
pyserial
pyqtgraph
ezdxf
scipy
zmq
